+++ Rectangle.h
@@ -29,7 +29,7 @@ class Rectangle : public EltGeo
 
 public:
 //----------------------------------------------------- Méthodes publiques
-	void Translater ( int dx, int dy );
+	void Translater ( long int dx, long int dy );
     // Mode d'emploi :
     //Appelle la méthode Point::Translater(dx,dy) des sommets p1 et p2 du rectangle
     // Contrat :
@@ -55,7 +55,7 @@ public:
     // Contrat :
     //
 
-    Rectangle ( int fx1,int fy1,int fx2,int fy2,string fname="",string fcommande="");
+    Rectangle ( long int fx1,long int fy1,long int fx2,long int fy2,string fname="",string fcommande="");
     // Mode d'emploi :
     //Ajoute les deux sommets du rectangle
     // Contrat :
